[Chore] #11 - CarouselSlider 적용
