+++ lib/screens/home.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:ttangkkeusmarket/widgets/AppBar.dart';
+import 'package:ttangkkeusmarket/widgets/CarouselList.dart';
 import 'package:ttangkkeusmarket/widgets/TabBar.dart';
-import 'package:ttangkkeusmarket/widgets/bottom_navbar.dart';
+import 'package:ttangkkeusmarket/widgets/BottomNavBar.dart';
+// import 'package:ttangkkeusmarket/widgets/HorizontalList.dart';
 
 class Main extends StatelessWidget {
   const Main({Key? key}) : super(key: key);
@@ -22,7 +24,7 @@ class HomeScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: BaseAppBar(appBar: AppBar(), title: "땅끗마켓", center: true),
-      body: const MainTabBar(),
+      body: MainTabBar()
     );
   }
 }
