+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   provider: ^6.0.2
+  google_fonts: ^2.3.2
 
 dev_dependencies:
   flutter_test:
@@ -48,6 +49,7 @@ dev_dependencies:
   flutter_lints: ^1.0.0
   flutter_native_splash: ^2.1.6
   flutter_launcher_icons: ^0.9.2
+  carousel_slider: ^4.1.1
 
 flutter_native_splash:
   color: "#ffffff"
@@ -96,19 +98,19 @@ flutter:
         - asset: assets/fonts/SF-Pro.ttf
     - family: NotoSans
       fonts:
-        - asset: assets/fonts/NotoSansCJKKr-Black.otf
+        - asset: assets/fonts/NotoSansKR-Black.otf
           weight: 900
-        - asset: assets/fonts/NotoSansCJKKr-Bold.otf
+        - asset: assets/fonts/NotoSansKR-Bold.otf
           weight: 700
-        - asset: assets/fonts/NotoSansCJKKr-Medium.otf
+        - asset: assets/fonts/NotoSansKR-Medium.otf
           weight: 500
-        - asset: assets/fonts/NotoSansCJKKr-Regular.otf
+        - asset: assets/fonts/NotoSansKR-Regular.otf
           weight: 400
         # - asset: assets/fonts/NotoSansCJKKr-DemiLight.otf
         #   weight: 350
-        - asset: assets/fonts/NotoSansCJKKr-Light.otf
+        - asset: assets/fonts/NotoSansKR-Light.otf
           weight: 300
-        - asset: assets/fonts/NotoSansCJKKr-Thin.otf
+        - asset: assets/fonts/NotoSansKR-Thin.otf
           weight: 100
   #
   # For details regarding fonts from package dependencies,
